+++ src/tutorial/lessons/index.ts
@@ -4,8 +4,9 @@ import { lesson04, lesson05, lesson06 } from "./module2";
 import { lesson07, lesson08, lesson09, lesson10, lesson11 } from "./module3";
 import { lesson12, lesson13 } from "./module4";
 import { lesson14 } from "./module5";
+import { lesson15 } from "./module6";
 
-export const LESSONS: Lesson[] = [lesson01, lesson02, lesson03, lesson04, lesson05, lesson06, lesson07, lesson08, lesson09, lesson10, lesson11, lesson12, lesson13, lesson14];
+export const LESSONS: Lesson[] = [lesson01, lesson02, lesson03, lesson04, lesson05, lesson06, lesson07, lesson08, lesson09, lesson10, lesson11, lesson12, lesson13, lesson14, lesson15];
 
 export const MODULES = [...new Set(LESSONS.map((l) => l.module))];